Using tabs instead of spaces
Some times in the life is good learning from your errors, and this is one of
them. It is better use tabs instead of spaces because you let to the
programmer decide about length of the tabs, so if you want select other you
can do it easilly, and your code will waste less space.